+++ lib/castle/context_merger.rb
@@ -2,12 +2,11 @@
 
 module Castle
   class ContextMerger
-    def initialize(context)
-      @main_context = Castle::Utils::Cloner.call(context)
-    end
-
-    def call(request_context)
-      Castle::Utils::Merger.call(@main_context, request_context)
+    class << self
+      def call(initial_context, request_context)
+        main_context = Castle::Utils::Cloner.call(initial_context)
+        Castle::Utils::Merger.call(main_context, request_context || {})
+      end
     end
   end
 end